back/requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.3
Flask-Migrate==4.0.4
psycopg2-binary==2.9.6
python-dotenv==1.0.0
instagrapi==1.16.2
